+++ src/subcommands/profile/mod.rs
@@ -51,7 +51,7 @@ pub async fn pick_minecraft_version() -> Result<String> {
 }
 
 /// Check that there isn't already a profile with the same name
-pub fn check_profile_name(config: &mut config::structs::Config, name: &str) -> Result<()> {
+pub fn check_profile_name(config: &config::structs::Config, name: &str) -> Result<()> {
     for profile in &config.profiles {
         if profile.name == name {
             bail!("A profile with name {} already exists", name);