+++ src/subcommands/profile/create.rs
@@ -1,27 +1,33 @@
+use crate::THEME;
+
 use super::{check_output_directory, check_profile_name, pick_minecraft_version};
 use anyhow::{bail, Result};
 use colored::Colorize;
 use dialoguer::{Confirm, Input, Select};
-use libium::{config, file_picker, misc};
+use libium::{
+    config::structs::{Config, ModLoader, Profile},
+    file_picker::pick_folder,
+    misc::get_minecraft_dir,
+};
 use std::path::PathBuf;
 
 #[allow(clippy::option_option)]
 pub async fn create(
-    config: &mut config::structs::Config,
+    config: &mut Config,
     import: Option<Option<String>>,
     game_version: Option<String>,
-    mod_loader: Option<config::structs::ModLoader>,
+    mod_loader: Option<ModLoader>,
     name: Option<String>,
     output_dir: Option<PathBuf>,
 ) -> Result<()> {
     let mut profile = match (game_version, mod_loader, name, output_dir) {
         (Some(game_version), Some(mod_loader), Some(name), output_dir) => {
             check_profile_name(config, &name)?;
-            let output_dir = output_dir.unwrap_or_else(|| misc::get_minecraft_dir().join("mods"));
+            let output_dir = output_dir.unwrap_or_else(|| get_minecraft_dir().join("mods"));
             if !output_dir.is_absolute() {
                 bail!("The provided output directory is not absolute, i.e. it is a relative path")
             }
-            config::structs::Profile {
+            Profile {
                 name,
                 output_dir,
                 game_version,
@@ -30,17 +36,16 @@ pub async fn create(
             }
         },
         (None, None, None, None) => {
-            let mut selected_mods_dir = misc::get_minecraft_dir().join("mods");
+            let mut selected_mods_dir = get_minecraft_dir().join("mods");
             println!(
                 "The default mods directory is {}",
                 selected_mods_dir.display()
             );
-            if Confirm::with_theme(&*crate::THEME)
+            if Confirm::with_theme(&*THEME)
                 .with_prompt("Would you like to specify a custom mods directory?")
                 .interact()?
             {
-                if let Some(dir) =
-                    file_picker::pick_folder(&selected_mods_dir, "Pick an output directory").await
+                if let Some(dir) = pick_folder(&selected_mods_dir, "Pick an output directory").await
                 {
                     check_output_directory(&dir).await?;
                     selected_mods_dir = dir;
@@ -48,7 +53,7 @@ pub async fn create(
             }
 
             let name = loop {
-                let name: String = Input::with_theme(&*crate::THEME)
+                let name: String = Input::with_theme(&*THEME)
                     .with_prompt("What should this profile be called?")
                     .interact_text()?;
 
@@ -69,7 +74,7 @@ pub async fn create(
 
             let selected_version = pick_minecraft_version().await?;
 
-            config::structs::Profile {
+            Profile {
                 name,
                 output_dir: selected_mods_dir,
                 mods: Vec::new(),
@@ -104,7 +109,7 @@ pub async fn create(
                     .iter()
                     .map(|profile| &profile.name)
                     .collect::<Vec<_>>();
-                Select::with_theme(&*crate::THEME)
+                Select::with_theme(&*THEME)
                     .with_prompt("Select which profile to import mods from")
                     .items(&profile_names)
                     .default(config.active_profile)